refactor: RequestHandler 클래스 리팩터링
- 엔드포인트별로 HTTP 요청을 처리할 로직 분리
